tests/wpt/web-platform-tests/speculation-rules/prerender/local-storage.html

<!DOCTYPE html>
<title>Same-origin prerendering can access localStorage</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const bc = new BroadcastChannel('prerender-channel');

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {once:true});
  });

  const initialValue = "initial_set";
  window.localStorage.setItem('initial', initialValue);

  // Start prerendering a page that attempts to access localStorage API.
  startPrerendering(`resources/local-storage-access.html`);
  const result = await gotMessage;
  assert_equals(
    result, initialValue,
      'prerendering page should be able to read from local storage');
  assert_equals(
      window.localStorage.getItem('prerender'), 'prerender_set',
      'prerendering page should be able to write to local storage');
}, 'prerendering page should be able to access local storage');

</script>
</body>
